Fixed raycast count data type

+++ Assets/Scripts/GameConstants.cs
@@ -6,7 +6,7 @@ namespace GameConstants
     public static class Constants
     {
         public const float MAX_RAYCAST_DISTANCE = 100;
-        public const float MAX_RAYCAST_COUNT = 100;
+        public const int MAX_RAYCAST_COUNT = 100;
     }
 
     namespace Enumerations
